Minor debugging

+++ verif/tests/actions/load_combinations/comb_container/test_comb_container_04.py
@@ -45,4 +45,5 @@ if comparisonOK:
    print('test '+fname+': ok.')
 else:
     lmsg.error(fname+' ERROR.')
-os.remove(texFileName) # Your garbage, you clean it.
+if(comparisonOK):
+    os.remove(texFileName) # Your garbage, you clean it.
